+++ server/plugins/local-mgmt-client.js
